requirements.txt
praw
requests
imgurpython
tqdm
coloredlogs
python-dotenv